requirements.txt
PyPDF2
pdfminer.six
pdfplumber
pdf2image
Pillow
pytesseract 
